net: ravb: Use pm_runtime_suspended() instead of pm_runtime_active()
The pm_runtime_active() is returns the value of:
(dev->power.runtime_status == RPM_ACTIVE || dev->power.disable_depth).
The runtime_status is set to active when the device is runtime resumed
and changed to RPM_SUSPENDED when the device is runtime suspended.
In the device registration phase the disabled_depth is set to 1 and the
runtime_status is set to RPM_SUSPENDED though the following call trace:

device_register()
  device_initialize()
    device_pm_init()
      pm_runtime_init()
        dev->power.runtime_status = RPM_SUSPENDED;
        dev->power.disable_depth = 1;

After that the disable_depth is decremented on pm_runtime_enable() and
incremented on pm_runtime_disable().
The runtime PM status is set to RPM_ACTIVE on runtime resume and
RPM_SUSPENDED on runtime suspend.

Requesting the interrupts in probe (when dev->power.disable_depth = 1)
w/o calling first pm_runtime_enable() leads to pm_runtime_active() check
in interrupt handlers to return true. If the interrupts are shared then
this leads to accessing hardware w/o it's runtime PM resources (clocks)
being enabled which in turn leads to synchronous aborts.

To solve this the pm_runtime_enable() was moved before requesting
interrupts.

If CONFIG_DEBUG_SHIRQ is enabled then interrupt handler is called on
free_irq(). The ravb interrupts are devres resources meaning that they
will be freed by devm_irq_release(). Call trace for calling IRQ handler
on free_irq() is as follows:

devm_irq_release()
  free_irq()
    __free_irq()
      // ..
      #ifdef CONFIG_DEBUG_SHIRQ
      if (action->flags & IRQF_SHARED) {
              local_irq_save(flags);
              action->handler(irq, dev_id);
              local_irq_restore(flags);
      }

This means that for runtime PM check to work in interrupt handler the
PM runtime enable needs to be devres resource as well. Otherwise we need to
call PM runtime disable on driver remove, this leads to
dev->power.disable_depth being incremented and pm_runtime_active() return
true.

To solve this the pm_runtime_enable() was changed to
devm_pm_runtime_enable().

With this, at the end of ravb_remove() function (and after PM runtime is
enabled in ravb_probe()) the runtime PM status of ravb device is
RPM_SUSPENDED and disable_depth = 0. This corresponds exactly to the
runtime PM suspend status. As this status is detected by
pm_runtime_suspended() function changed the pm_runtime_active() calls
to pm_runtime_suspended().

Fixes: 32f012b8c01c ("net: ravb: Move getting/requesting IRQs in the probe() method")
Fixes: bbf2345fa658 ("net: ravb: Return cached statistics if the interface is down")
Fixes: a71a50e391bf ("net: ravb: Do not apply features to hardware if the interface is down")
Signed-off-by: Claudiu Beznea <claudiu.beznea.uj@bp.renesas.com>