reset: rzg2l-usbphy-ctrl: Add support to power on/off the USB area
Before accessing the USB area of the RZ/G3S SoC the PWRRDY bit of the
SYS_USB_PWRRDY register need to be cleared. When USB area is not used the
PWRRDY bit of the SYS_USB_PWRRDY register need to be set. This register is
in the SYSC controller address space and it is exposed to consummers
though a PM domain.

The USB modules available on the RZ/G3S SoC that need this bit set are:
- USB ch0 (supporting host and peripheral mode)
- USB ch2 (supporting host mode)
- USBPHY control

As the USBPHY control is the root device for all the other USB channels
(USB ch0, USB ch1) add support to set the PWRRDY for the USB area when
initializing the USBPHY control. The driver is getting reference to the
power domain and powers it on.

Signed-off-by: Claudiu Beznea <claudiu.beznea.uj@bp.renesas.com>