soc: renesas: Add SYSC driver for RZ/G3S
The RZ/G3S SYS Controller has 2 registers (one for PCI one for USB) that
need to be configured on before/after powering off/on the PCI or USB
ares. For this the added SYSC controller driver registers and auxiliary
device which controlls the PCI/USB behavior though power domains.

The auxiliary device infrastructure was chosed to allow extending the
SYSC controller driver with other functionalities (e.g., SoC
Identification).

Signed-off-by: Claudiu Beznea <claudiu.beznea.uj@bp.renesas.com>